Fixed Issue #1, #4 & #5
