+++ Handlers/DataHandler.cs
@@ -1,4 +1,5 @@
-﻿using Newtonsoft.Json;
+﻿using MBModManager.Data;
+using Newtonsoft.Json;
 using System.IO;
 
 namespace MBModManager.Handlers
