--- bin/Debug/net6.0-windows/settings.json
@@ -1,10 +1,0 @@
-{
-  "GamePath": "C:\\Program Files (x86)\\Steam\\steamapps\\common\\MonBazou",
-  "GameVersion": "0.0.0",
-  "SavesPath": "C:\\Users\\Drick\\AppData\\LocalLow\\Santa Goat\\Mon Bazou",
-  "InternalData": [
-    "v1.0.0",
-    "4f78160a-6f22-4d53-94ce-afd993ffd6e5",
-    "https://github.com/BepInEx/BepInEx/releases/download/v5.4.21/BepInEx_x64_5.4.21.0.zip"
-  ]
-}
